+++ src/main/java/chess/controller/mapper/response/ChessBoardStateFormatter.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-public class ChessBoardStateMapper {
+public class ChessBoardStateFormatter {
     private static final String EMPTY_PIECE_SYMBOL = ".";
 
     public static List<List<String>> convertToConsoleViewBoard(Map<Position, Piece> piecesPosition) {
